Move tooltip mixin to the client package, use localization for End Crystal tooltip
