+++ src/client/java/com/lgmrszd/anshar/mixin/client/ItemMixin.java
@@ -1,0 +1,21 @@
+package com.lgmrszd.anshar.mixin.client;
+
+import net.minecraft.client.item.TooltipContext;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.text.Text;
+import net.minecraft.world.World;
+import org.jetbrains.annotations.Nullable;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+import java.util.List;
+
+@Mixin(Item.class)
+public class ItemMixin {
+    @Inject(method = "appendTooltip", at = @At("HEAD"))
+    public void anshar$addToTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context, CallbackInfo ci) {
+    }
+}
