+++ src/client/resources/anshar.client.mixins.json
@@ -1,15 +1,17 @@
 {
-	"required": true,
-	"package": "com.lgmrszd.anshar.mixin.client",
-	"compatibilityLevel": "JAVA_17",
-	"client": [
-		"AbstractClientPlayerEntityMixin",
-		"BackgroundRendererMixin",
-		"WorldRendererMixin",
-		"InGameHudMixin",
-		"HeldItemRendererMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
+  "required": true,
+  "package": "com.lgmrszd.anshar.mixin.client",
+  "compatibilityLevel": "JAVA_17",
+  "client": [
+    "AbstractClientPlayerEntityMixin",
+    "BackgroundRendererMixin",
+    "WorldRendererMixin",
+    "InGameHudMixin",
+    "HeldItemRendererMixin",
+    "ItemMixin",
+    "EndCrystalItemMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
 }
