+++ src/main/java/com/lgmrszd/anshar/mixin/EndCrystalItemMixin.java
@@ -7,14 +7,10 @@ import com.lgmrszd.anshar.beacon.EndCrystalComponent;
 import com.lgmrszd.anshar.beacon.EndCrystalItemContainer;
 import com.lgmrszd.anshar.beacon.IEndCrystalComponent;
 import com.lgmrszd.anshar.frequency.NetworkManagerComponent;
-import net.minecraft.client.item.TooltipContext;
-import net.minecraft.item.ItemStack;
-import net.minecraft.text.Text;
-import org.jetbrains.annotations.Nullable;
+import net.minecraft.item.Item;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 import net.minecraft.block.BlockState;
@@ -33,7 +29,11 @@ import net.minecraft.world.event.GameEvent;
 import static com.lgmrszd.anshar.beacon.EndCrystalComponent.MAX_DISTANCE;
 
 @Mixin(EndCrystalItem.class)
-public abstract class EndCrystalItemMixin extends ItemMixin {
+public abstract class EndCrystalItemMixin extends Item {
+
+    public EndCrystalItemMixin(Settings settings) {
+        super(settings);
+    }
 
     @Inject(method = "useOnBlock", at = @At("HEAD"), cancellable = true)
     public void anshar$useOnBlock(ItemUsageContext context, CallbackInfoReturnable<ActionResult> ci){
@@ -79,17 +79,5 @@ public abstract class EndCrystalItemMixin extends ItemMixin {
             ci.setReturnValue(ActionResult.success(world.isClient));
             ci.cancel();
         }
-    }
-
-    @Override
-    public void anshar$addToTooltip(ItemStack stack, @Nullable World world, List<Text> tooltip, TooltipContext context, CallbackInfo ci) {
-        EndCrystalItemContainer container = ModApi.END_CRYSTAL_ITEM.find(stack, null);
-        // TODO make prettier
-        container.getBeaconPos().ifPresent(pos -> {
-            tooltip.add(Text.literal(String.format(
-                    "Stored position: %s",
-                    pos
-            )));
-        });
     }
 }
