+++ src/main/resources/assets/anshar/lang/en_us.json
@@ -1,4 +1,7 @@
 {
     "anshar.help.transport.gate": "Hold forward to enter Star Gate",
-    "anshar.help.transport.exit": "Crouch to exit"
+    "anshar.help.transport.exit": "Crouch to exit",
+
+    "anshar.tooltip.end_crystal.linked": "Linked to the Beacon at: %d %d %d",
+    "anshar.tooltip.end_crystal.linked.distance": "(Distance: %s)"
 }
